pyproject.toml
[project]
name = "batch-inference"
version = "0.1.0"
description = "AWS Bedrock batch inference SDK with boto3-compatible invoke_model API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioboto3>=13.0.0",
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "python-dotenv>=1.0.0",
    "tqdm>=4.67.1",
]
